fix: Importe de bibliotecas corretas e alterações no makefile para compilar
