+++ src/modules/appointment/usecases.c
@@ -1,6 +1,9 @@
+#include "usecases.h"
 #include <stdio.h>
 #include <stdbool.h>
-#include "./../utils/interfaces.h"
+#include <string.h>
+#include <stdlib.h>
+#include "./../../utils/interfaces.h"
 
 #ifdef __unix__
 
@@ -8,7 +11,6 @@
 #include <unistd.h>
 
 #endif
-
 
 void showAppointmentMenu() {
     #ifdef __unix__
