+++ src/modules/client/usecases.h
@@ -1,0 +1,1 @@
+void showClientMenu(void);
