+++ src/modules/office/usecases.h
@@ -1,0 +1,1 @@
+void showOfficeMenu(void);
