+++ src/utils/interfaces.c
@@ -3,8 +3,10 @@
 #include <string.h>
 #include <stdlib.h>
 #include <stdbool.h>
-#include "./../appointment/usecases.h"
+#include "../modules/appointment/usecases.h"
 #include "./../lawyer/usecases.h"
+#include "./../office/usecases.h"
+#include "./../client/usecases.h"
 
 #ifdef __unix__
 
@@ -237,122 +239,6 @@ void showMainMenu() {
             #endif
             isSelected = false;
             if (option >= 0 && option <= 5) actions[option](); else loop = false;
-        }
-    }
-}
-
-void showClientMenu() {
-    #ifdef __unix__
-        struct termios originalTerminal;
-        tcgetattr(STDIN_FILENO, &originalTerminal);
-    #endif
-    int aux, option = 0, size = 5;
-    bool isSelected = false, loop = true;
-    char optionsStyles[size][11];
-    setOptionsStyle(optionsStyles, size);
-    while (loop) {
-        #ifdef __unix__
-            system("clear");
-            enableRawMode();
-        #else
-            system("cls");
-        #endif
-        if (!isSelected) {
-            printf("----- Menu Cliente -----\n");
-            printf("|                      |\n");
-            printf("| %s1. Cadastar cliente%s  |\n", optionsStyles[0], RESET_STYLE);
-            printf("| %s2. Mostar cliente%s    |\n", optionsStyles[1], RESET_STYLE);
-            printf("| %s3. Editar cliente%s    |\n", optionsStyles[2], RESET_STYLE);
-            printf("| %s4. Exluir cliente%s    |\n", optionsStyles[3], RESET_STYLE);
-            printf("| %s5. Voltar ao menu%s    |\n", optionsStyles[4], RESET_STYLE);
-            printf("|                      |\n");
-            printf("------------------------\n");
-            
-            aux = option;
-            selectOption(&option, size - 1, &isSelected);
-
-            strcpy(optionsStyles[aux], RESET_STYLE);
-            strcpy(optionsStyles[option], CYAN_UNDERLINE_TEXT);
-        } else {
-            #ifdef __unix__
-                disableRawMode(&originalTerminal);
-            #endif
-            isSelected = false;
-            switch (option) {
-                case 0:
-                    //createClient();
-                    break;
-                case 1:
-                    //readClient();
-                    break;
-                case 2:
-                    //updateClient();
-                    break;
-                case 3:
-                    //deleteClient();
-                    break;
-                default:
-                    loop = false;
-                    break;
-            }
-        }
-    }
-}
-
-void showOfficeMenu() {
-    #ifdef __unix__
-        struct termios originalTerminal;
-        tcgetattr(STDIN_FILENO, &originalTerminal);
-    #endif
-    int aux, option = 0, size = 5;
-    bool isSelected = false, loop = true;
-    char optionsStyles[size][11];
-    setOptionsStyle(optionsStyles, size);
-    while (loop) {
-        #ifdef __unix__
-            system("clear");
-            enableRawMode();
-        #else
-            system("cls");
-        #endif
-        if (!isSelected) {
-            printf("----- Menu Escritório -----\n");
-            printf("|                         |\n");
-            printf("| %s1. Cadastar escritório%s  |\n", optionsStyles[0], RESET_STYLE);
-            printf("| %s2. Mostar escritório%s    |\n", optionsStyles[1], RESET_STYLE);
-            printf("| %s3. Editar escritório%s    |\n", optionsStyles[2], RESET_STYLE);
-            printf("| %s4. Exluir escritório%s    |\n", optionsStyles[3], RESET_STYLE);
-            printf("| %s5. Voltar ao menu%s       |\n", optionsStyles[4], RESET_STYLE);
-            printf("|                         |\n");
-            printf("---------------------------\n");
-            
-            aux = option;
-            selectOption(&option, size - 1, &isSelected);
-
-            strcpy(optionsStyles[aux], RESET_STYLE);
-            strcpy(optionsStyles[option], CYAN_UNDERLINE_TEXT);
-        } else {
-            #ifdef __unix__
-                disableRawMode(&originalTerminal);
-            #endif
-            isSelected = false;
-            switch (option) {
-                case 0:
-                    //createOffice();
-                    break;
-                case 1:
-                    //readOffice();
-                    break;
-                case 2:
-                    //updateOffice();
-                    break;
-                case 3:
-                    //deleteOffice();
-                    break;
-                default:
-                    loop = false;
-                    break;
-            }
         }
     }
 }
